+++ mtp/ops_direct.go
@@ -6,37 +6,7 @@ import (
 	"io"
 	"log"
 	"math/rand"
-	"time"
 )
-
-func init() {
-	rand.Seed(time.Now().UnixNano())
-}
-
-func (d *DeviceGoUSB) OpenSession() error {
-	if d.session != nil {
-		return fmt.Errorf("session already open")
-	}
-
-	var req, rep Container
-	req.Code = OC_OpenSession
-
-	// avoid 0xFFFFFFFF and 0x00000000 for session IDs.
-	sid := uint32(rand.Int31()) | 1
-	req.Param = []uint32{sid} // session
-
-	// If opening the session fails, we want to be able to reset
-	// the device, so don't do sanity checks afterwards.
-	if err := d.runTransaction(&req, &rep, nil, nil, 0); err != nil {
-		return err
-	}
-
-	d.session = &sessionData{
-		tid: 1,
-		sid: sid,
-	}
-	return nil
-}
 
 // OpenSession opens a session, which is necesary for any command that
 // queries or modifies storage. It is an error to open a session
@@ -66,33 +36,11 @@ func (d *DeviceDirect) OpenSession() error {
 }
 
 // Closes a sessions. This is done automatically if the device is closed.
-func (d *DeviceGoUSB) CloseSession() error {
-	var req, rep Container
-	req.Code = OC_CloseSession
-	err := d.RunTransaction(&req, &rep, nil, nil, 0)
-	d.session = nil
-	return err
-}
-
-// Closes a sessions. This is done automatically if the device is closed.
 func (d *DeviceDirect) CloseSession() error {
 	var req, rep Container
 	req.Code = OC_CloseSession
 	err := d.RunTransaction(&req, &rep, nil, nil, 0)
 	d.session = nil
-	return err
-}
-
-func (d *DeviceGoUSB) GetData(req *Container, info interface{}) error {
-	var buf bytes.Buffer
-	var rep Container
-	if err := d.RunTransaction(req, &rep, &buf, nil, 0); err != nil {
-		return err
-	}
-	err := Decode(&buf, info)
-	if d.Debug.MTP && err == nil {
-		d.log.WithField("prefix", "mtp").Debugf("MTP decoded %#v", info)
-	}
 	return err
 }
 
@@ -107,12 +55,6 @@ func (d *DeviceDirect) GetData(req *Container, info interface{}) error {
 		log.Printf("MTP decoded %#v", info)
 	}
 	return err
-}
-
-func (d *DeviceGoUSB) GetDeviceInfo(info *DeviceInfo) error {
-	var req Container
-	req.Code = OC_GetDeviceInfo
-	return d.GetData(&req, info)
 }
 
 func (d *DeviceDirect) GetDeviceInfo(info *DeviceInfo) error {
@@ -179,13 +121,6 @@ func (d *DeviceDirect) SetDevicePropValue(propCode uint32, src interface{}) erro
 	req.Code = OC_SetDevicePropValue
 	req.Param = []uint32{propCode}
 	return d.SendData(&req, &rep, src)
-}
-
-func (d *DeviceGoUSB) GetDevicePropValue(propCode uint32, dest interface{}) error {
-	var req Container
-	req.Code = OC_GetDevicePropValue
-	req.Param = []uint32{propCode}
-	return d.GetData(&req, dest)
 }
 
 func (d *DeviceDirect) GetDevicePropValue(propCode uint32, dest interface{}) error {
